DBZ-2363 Extracting separate test class for logical decoding message tests
